requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
joblib
lightgbm
sentence-transformers
faiss-cpu
